Move pythod fallbacks back to base_tasks.py
